+++ app/src/main/java/com/github/droidworksstudio/mlauncher/ui/SettingsFragment.kt
@@ -363,6 +363,14 @@ class SettingsFragment : Fragment() {
                             state = remember { mutableStateOf(prefs.useCustomIconFont) },
                         ) { toggleCustomIconFont() }
                     },
+                    { _, onChange ->
+                        SettingsToggle(
+                            title = stringResource(R.string.using_feature_phone),
+                            fontSize = iconFs,
+                            onChange = onChange,
+                            state = remember { mutableStateOf(prefs.usingFeaturePhone) },
+                        ) { toggleFeaturePhone() }
+                    },
                 )
             )
             SettingsArea(
@@ -809,6 +817,11 @@ class SettingsFragment : Fragment() {
     private fun toggleCustomIconFont() {
         prefs.useCustomIconFont = !prefs.useCustomIconFont
     }
+
+    private fun toggleFeaturePhone() {
+        prefs.usingFeaturePhone = !prefs.usingFeaturePhone
+    }
+
 
     private fun toggleAllAppsText() {
         prefs.useAllAppsText = !prefs.useAllAppsText